Remote access, fixes remote user session issue
