+++ mxcube3/ui/main.jsx
@@ -80,7 +80,7 @@ export default class App extends React.Component {
   componentWillMount() {
     const persistor = persistStore(store,
            { blacklist: ['remoteAccess', 'beamline', 'sampleChanger',
-                         'form', 'login', 'general', 'logger', 'points'],
+                         'form', 'login', 'general', 'logger', 'points', 'queue'],
              storage: new ServerStorage() },
              () => {
                /* eslint-disable react/no-set-state */
